DABOM-482 rename: global -> common 변경
