+++ src/main/java/com/template/worker/common/config/SchedulingConfig.java
@@ -1,4 +1,4 @@
-package com.template.worker.global.config;
+package com.template.worker.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
