+++ src/main/java/com/template/worker/common/listener/JobLogger.java
@@ -1,4 +1,4 @@
-package com.template.worker.global.listener;
+package com.template.worker.common.listener;
 
 import org.springframework.stereotype.Component;
