+++ src/main/java/com/template/worker/jobs/reconciliation/job/DbRedisReconciliationJobConfig.java
@@ -6,8 +6,8 @@ import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import com.template.worker.global.listener.JobFailureAlertListener;
-import com.template.worker.global.listener.JobResultListener;
+import com.template.worker.common.listener.JobFailureAlertListener;
+import com.template.worker.common.listener.JobResultListener;
 import com.template.worker.jobs.common.support.BatchJobConstants;
 import com.template.worker.jobs.reconciliation.listener.DbRedisReconciliationJobListener;
 import com.template.worker.jobs.reconciliation.step.AcquireReconciliationLockStepConfig;
