+++ src/test/java/com/template/worker/common/launcher/BatchJobLauncherTest.java
@@ -1,4 +1,4 @@
-package com.template.worker.global.launcher;
+package com.template.worker.common.launcher;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
