+++ src/test/java/com/template/worker/jobs/usagereset/writer/CustomerMonthlyUsageResetWriterTest.java
@@ -25,7 +25,7 @@ import org.springframework.data.redis.connection.StringRedisConnection;
 import org.springframework.data.redis.core.RedisCallback;
 import org.springframework.data.redis.core.StringRedisTemplate;
 
-import com.template.worker.global.util.RedisKeyGenerator;
+import com.template.worker.common.util.RedisKeyGenerator;
 import com.template.worker.jobs.usagereset.model.FamilyMemberUsageResetTarget;
 import com.template.worker.jobs.usagereset.support.MonthlyUsageResetJobConstants;
 import com.template.worker.jobs.usagereset.support.MonthlyUsageResetJobParameterSupport;
